refactor(sensor_values_page): extracted system microphone chart into it's own file
